Added Support for Syncronous Testing of Promise Presence
While the resolution of a promise is clearly an asynchronous thing, testing for the presence of a promise is not, so it should be possible to test if a given value is a promise using `validate()` or `validate.single()`.

This very small commit makes that possible by adding an option `allowPromises` to `validate()`.